requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
elasticsearch==7.17.10
pandas==2.1.1
numpy==1.24.3
python-dotenv==1.0.0
gunicorn==21.2.0
celery==5.3.1
redis==4.6.0
werkzeug==2.3.7
python-dateutil==2.8.2
jieba==0.42.1 